fix: focus issues
